# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DEV-001 Thyroid Pre-Visit Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patient-auth/verify-dob:
    post:
      operationId: verifyDob
      tags: [patient-auth]
      summary: Step 1 — verify date of birth against magic link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DobVerification"
      responses:
        "200":
          description: DOB matched — OTP sent to patient phone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DobVerificationResult"
        "400":
          description: Wrong DOB / rate-limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patient-auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [patient-auth]
      summary: Step 2 — verify OTP and issue patient session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerification"
      responses:
        "200":
          description: OTP matched — patient session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientSession"
        "400":
          description: Wrong OTP / rate-limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Appointments ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    patientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ────────────────────────────────────────────────────────────────
...
    DobVerification:
      type: object
      required: [token, dateOfBirth]
      properties:
        token:
          type: string
        dateOfBirth:
          type: string
          description: "YYYY-MM-DD"
    DobVerificationResult:
      type: object
      required: [otpSent]
      properties:
        otpSent:
          type: boolean
        phone:
          type: ["string", "null"]
          description: Masked phone number
        sessionId:
          type: ["string", "null"]
    OtpVerification:
      type: object
      required: [token, otp]
      properties:
        token:
          type: string
        otp:
          type: string
    PatientSession:
      type: object
      required: [patientToken, appointmentId]
      properties:
        patientToken:
          type: string
        appointmentId:
          type: string

    # ── Appointments ────────────────────────────────────────────────────────
...